Migrated themes to its relative file
